backend/gateway-service/requirements.txt
fastapi>=0.116.1
httpx>=0.28.1
pydantic[email]>=2.11.7
pydantic-settings>=2.3.4
python-dotenv>=1.0.1
sqlalchemy>=2.0.42
uvicorn[standard]>=0.35.0
